+++ src/main/java/com/ikomat/baseball/CombinationPiches/domain/repositories/Dao.java
@@ -5,5 +5,5 @@ import org.springframework.data.jpa.repository.JpaRepository;
 import org.springframework.stereotype.Repository;
 
 @Repository
-public interface Dao extends JpaRepository<MatchInfo,Integer> {
+public interface Dao extends JpaRepository<MatchInfo,Long> {
 }
