+++ src/main/java/com/ikomat/baseball/CombinationPiches/domain/repositories/DaoSituationEntity.java
@@ -5,5 +5,5 @@ import org.springframework.data.jpa.repository.JpaRepository;
 import org.springframework.stereotype.Repository;
 
 @Repository
-public interface DaoSituationEntity extends JpaRepository<SituationEntity,Integer> {
+public interface DaoSituationEntity extends JpaRepository<SituationEntity,Long> {
 }
